Balance

+++ src/Education.Client/Gateways/Models/IGameCurrency.cs
@@ -4,6 +4,6 @@ public interface IGameCurrency
 {
 }
 
-public sealed record ExperienceCurrency(uint Value) : IGameCurrency;
+public sealed record ExperienceCurrency(ulong Value) : IGameCurrency;
 
-public sealed record SilverCurrency(int Value) : IGameCurrency;
+public sealed record SilverCurrency(long Value) : IGameCurrency;
